Created Host class which will be used to start Host functions
Created Host class. At the moment this class contains barebone information
gathering about the Host, but will be further developed to do all the
functions we need for hosts.

The Rupert file has also been modified to provide the 'available_connection'
and 'open' methods, which I use to reduce the amount of code I have to type
for tests.