Added support for editing a virtual machine.
Added support for editing a virtual machine. When you go to edit a virtual
machine, the libvirt library expects a UUID identifier to assume that you are
in fact updating a machine. This UUID is automagically provided by libvirt
when the virtual machine is created. If no UUID is provided, then libvirt
assumes you are trying to define a new domain.